Estoc Tweaks
The estoc is currently a very peculiar weapon.
It has a chop, despite being needle thin. It has an odd setup where next to someone, it's on par with a spear just about. The entire point of something like an estoc is that it's long and designed to deal with armour, so I'd gone ahead and slapped some numbers about.

Gave it a unique lunge with AP to allow it to bleed through armour. Not as great as the actual adjacent stab. In fact, it's only useful for bleeding. But that's the point of something like the estoc, I'd argue.

Additionally, I'd removed both one handed and two-handed chopping. Gave it the standard sword thrust when one-handed, and for two-handed I'd given it the greater sword peeling. Which, mind, I'd argue is thematic. It's five strikes to peel in that case, however. One more than standard sword peeling. We can drop the peeling altogether, if this is a PR we even want.

Further, I'd removed the swing delay on both the lunge and thrust, putting it on par with the spear for attack speed and click cooldown. Is this a good idea? I've no clue. I'd just like to see anti-armour variety, personally.

I was just super disappointed that the estoc is effectively a greatsword, with less damage, and a weird adjacent pen. But worse aside from that.

I miss the old estoc...