Licker Bat Removal | Wretch PQ Lowering
Licker bat form removed outright. Xray vision. Immunity to silver. Could ignore the sun. Miracles did nothing. Did not inherit biotypes of parent mob. Among a whole host of many other issues. It was nearly as bad as OldZad form. Maybe worse, honestly, since it facilitates a super powerful stupid antag to do super powerful stupid antag stuff.

Additionally, it drops Wretch PQ requirement to 30. We've levelled out in terms of pop. Not a lot of new friends picking it for the sake of fragging, and this is higher than it was anyways previously. By a tiny bit. Pipeline was 60, lowered by a contributor to 50, now at 30. We'll re-raise / adjust as needed.