Monk Loadouts
Simply separates the monk setups into loadouts. We'd people who wanted dodge expert staves, for some reason. Others who want crit resist punching, for obvious reasons. Etc. Whatever. This was requested by a few people, and it's such a simple thing with so few people touching it, that it shouldn't be a big deal.